refactor: Add registration route to regController
This commit adds the registration route to the `regController.js` file. The `register` function is implemented to handle the registration process by calling the `registerUser` function from the `regService` module. If successful, the response is sent as JSON. If an error occurs, a 500 status code is returned with an error message.